fixed lat long datatypes

+++ backend/db/migrations/20240402153611-create-spot.js
@@ -41,10 +41,10 @@ module.exports = {
           allowNull: false,
         },
         lat: {
-          type: Sequelize.INTEGER,
+          type: Sequelize.DECIMAL,
         },
         lng: {
-          type: Sequelize.INTEGER,
+          type: Sequelize.DECIMAL,
         },
         name: {
           type: Sequelize.STRING,
